Extracted subcommand from signing engine
