+++ ethsigner/app/build.gradle
@@ -29,11 +29,7 @@ dependencies {
 
   implementation project(':ethsigner:core')
   implementation project(':ethsigner:signing-api')
-  implementation project(':ethsigner:signer:hashicorp')
-  implementation project(':ethsigner:signer:file-based')
-  implementation project(':ethsigner:signer:multikey')
-  implementation project(':ethsigner:signer:azure')
-  implementation project(':ethsigner:signer:raw')
+  implementation project(':ethsigner:subcommands')
   implementation project(':ethsigner:commandline')
 
   implementation 'com.google.guava:guava'
