+++ ethsigner/subcommands/src/main/java/tech/pegasys/ethsigner/subcommands/FileBasedSubCommand.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.ethsigner.signer.filebased;
+package tech.pegasys.ethsigner.subcommands;
 
 import static tech.pegasys.ethsigner.DefaultCommandValues.MANDATORY_FILE_FORMAT_HELP;
 
@@ -27,6 +27,7 @@ import picocli.CommandLine;
 import picocli.CommandLine.Command;
 import picocli.CommandLine.Option;
 import picocli.CommandLine.Spec;
+import tech.pegasys.ethsigner.signer.filebased.FileBasedSignerFactory;
 
 /** File-based authentication related sub-command */
 @Command(
