+++ settings.gradle
@@ -20,5 +20,5 @@ include 'ethsigner:signer:hashicorp'
 include 'ethsigner:signer:file-based'
 include 'ethsigner:signer:azure'
 include 'ethsigner:signer:multikey'
-include 'ethsigner:signer:raw'
+include 'ethsigner:subcommands'
 include 'ethsigner:commandline'
